backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.6
pydantic-settings==2.8.1
langchain==0.3.20
langchain-openai==0.3.7
langchain-chroma==0.2.2
chromadb==0.6.3
openai==1.66.3
pytesseract==0.3.13
pillow==11.1.0
python-dotenv==1.0.1
sqlalchemy==2.0.39
psycopg[binary]==3.2.13
asyncpg==0.30.0
